fix depreciation warning in G2indx.findMV fix numpy round bug in G2lattice.MaxIndex fix bug in G2pwdGUI.OnFindMV fix bug in G2spc.SSpcGroup
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@4810 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ GSASIIindex.py
@@ -272,7 +272,7 @@ def findMV(peaks,controls,ssopt,Inst,dlg):
                 i += 1
             else:
                 Vec.append(vec[j])
-        return np.array(Vec)  
+        return np.array(Vec,dtype=float)  
      
     def ZSSfunc(values,peaks,dmin,Inst,SGData,SSGData,vec,Vref,maxH,A,wave,Z,dlg=None):
         Vec = Val2Vec(vec,Vref,values)
